telcopulse_dashboard/requirements.txt
# streamlit
# boto3
# pandas
# pyathena
streamlit==1.45.0
pandas==2.2.1
pyarrow==20.0.0
boto3==1.34.69
python-dotenv==1.0.1 
plotly==6.0.1
fastparquet
s3fs==2025.3.2